chore(browserstack): add concurrency field
this commit adds a concurrency string, 'run_browserstack_queue' to the
'run_browserstack' job in ci.

this prevents a condition in ci where two simultaneous browserstack runs
are attempted at the same time. stencil's browserstack license only
allows for five parallel runs at a given time. as of the time of this
writing, stencil uses three browsers per run of browserstack. as a
result, two runs of the job will result in requesting more browsers than
our license permits. on the browserstack side, the extra run (number six
of the five) is queued to run once one of the five completes. however,
on the karma side things, we don't have a great way to check for these
queued jobs and poll until one is done. as a result, we experience
timeouts/disconnects with browserstack as a result of waiting for a
browser instance (that's queued) to start up. oftentimes, we hit a
jasmine/karma timeout of ~30_000 milliseconds.

with this change, we are able to continue to have the flexability to
re-run just the browserstack tests, while blocking them from running
when we aren't sure that we can acquire all the necessary instances. it
should be noted that this does not handle cases where a developer is
tunneling into browserstack from their local machine

the name for the concurrency field was chosen to:
- make it easily idenifiable as a part of running browswerstack
- not overlap with the name of the run_browserstack job completely to
  avoid any confusion when deugging in the future