refactor(karma): document karma utilities (#3453)
this commit lightly refactors the various karma utilities used by
browserstack tests. the motivation for this effort was STENCIL-433,
wherein we are attempting to stabilize the browserstack tests. there
was an initial belief that some of the flakiness could have been a
result of the contents of these utils. that is not longer suspected -
however, the code changes found in this commit are (subjectively) nice
cleanups to have for future maintainers

STENCIL-433: Stabilize Browserstack Tests