chore(github): split browserstack into two jobs (#3454)
this commit takes the pre-existing browserstack workflow (renamed from
`browserstack.yml` to `test-browserstack.yml`) and splits one job into two:
one for building stencil, and another to build the downstream applications
+ run browserstack

the intent behind this commit is to make re-running browserstack a
faster operation for stencil engineers. rather than waiting
approximately 8 minutes to see if a job passes, we take the flakier of
the two (running browserstack) and move it into its own job. from the
github ui, we can "rerun failed jobs" for this workflow, resulting in a
~3.5 minute turnaround time. this is a part of a larger effort to
stabilize browserstack tests. while this commit doesn't perform any
stabilizations per se, it does help mitigate the time lost due to flaky
tests.

the first job makes use of our pre-existing `build_core` job, which
should meet our requirements for preventing secrets from leaking.
we do still end up with a double build (one in the main ci workflow,
and one here). the good news is that the two builds/caches are isolated
from one another, which means for now we don't have to worry about
race conditions between the Browserstack build and the main one. In the
future, we'd like to see if there's anything that can be done in this regard
(either using the same build step or smarter caching), but at the least
for now this PR moves the needle a little bit closer in the direction of ci
stability (or having a better experience when its flaky)

STENCIL-433 Stabilize Browserstack CI Tests