chore(compiler): make outputTargets required on the ValidatedConfig type
This makes the `outputTargets` field required on the `ValidatedConfig`
type, allowing us to freely access `.outputTargets` on a validated
configuration object in the compiler code without having to use the `!`
operator, check `=== undefined`, etc.

This includes changes to the places where a `ValidatedConfig` object is
created to default `.outputTargets` to `[]` if no suitable value is
present.