chore(): re-remove node 12 from browserstack
this commit re-removes node 12 from runs of browserstack in favor of
using node 14. this was previously done in
2d78e68f0347a76773c18d81b9d562f1849ff708 as as part of #3344. however,
in b9fbe85b8549d23fcc2dd93991c44135feb6428a (#3454), the original file
was deleted in favor of `test-browserstack.yml`. following merging that
latter commit into the v3.0.0-dev branch, this changeset ports the
original change over to the new file