chore(browserstack): re-remove node 12 (#3478)
this commit re-removes node 12 from runs of browserstack in favor of
using node 14. this was previously done in
0749dc088f47f142eda09975c4e1b214abcdf84a as as part of #3342. however,
in b9fbe85b8549d23fcc2dd93991c44135feb6428a (#3454), the original file
was deleted in favor of `test-browserstack.yml`. following merging that
latter commit into the v3.0.0-dev branch, this changeset ports the
original change over to the new file