chore(config): use ValidatedConfig in hasAppTarget
this commit updates `telemetry.ts#hasAppTarget` to use `ValidatedConfig`
as its argument, rather than an unvalidated one. the single caller of
this function (outside of tests) already has been converted to receive a
`ValidatedConfig`, making this conversion a matter of only updating the
function signature.

tests for this function were updated to adhere to the stricter
requirements of a `ValidatedConfig`