chore(mocks): update mockConfig, mockValidatedConfig to accept any property (#3485)
this commit adds a new parameter to `mockValidatedConfig`, `overrides`.
the argument is optional, and defaults to an empty object literal. it is
spread over the returned object to override and defaults put in place by
this method. the `sys` argument is also removed, as it now can be safely
derived from the `overrides` argument (and still falls back to a new
`TestingSystem` if `overrides.sys` is falsy

this commit adds an `overrides` parameter to `mockConfig`. `overrides`
is a partial instance of `UnvalidatedConfig`, which defaults to an empty
object literal. its contents are spread over the returned object,
overriding the values put in place by default. it also removes the `sys` arg.
' this value is now derived from the provided `overrides`, falling back to a
`TestingSystem` otherwise.

the properties on `mockConfig` are sorted in this commit, to make
finding them (by a human) easier. it is assumed that this is the "best"
ordering for them, as no other grouping seems "obvious" at this time

STENCIL-486: Update mockConfig, mockValidatedConfig to accept any property