chore(config): propagate ValidatedConfig through telemetry directory (#3484)
this commit updates `telemetry.ts#hasAppTarget` to use `ValidatedConfig`
as its argument, rather than an unvalidated one. the single caller of
this function (outside of tests) already has been converted to receive a
`ValidatedConfig`, making this conversion a matter of only updating the
function signature.

tests for this function were updated to adhere to the stricter
requirements of a `ValidatedConfig`

this commit updates `telemetry.ts#getActiveTargets` to use
`ValidatedConfig` as its argument, rather than an unvalidated one. the
single caller of this function already has been converted to receive a
`ValidatedConfig`, making this conversion a matter of only updating the
function signature.

this commit updates the aforementioned function to use `ValidatedConfig`
as its argument, rather than an unvalidated one. the single caller of
this function (outside of tests) already has been converted to receive a
`ValidatedConfig`, making this conversion a matter of only updating the
function signature.

tests for this function were updated to adhere to the stricter
requirements of a `ValidatedConfig`. because we now use a stricter
object, it has additional fields on it. this broke the existing
assertions on the function's tests. the tests have been updated to
maintain the original spirit/intent of the assertions.

`d.Config` was kept as the return type (and consequently the type in
`CONFIG_PROPS_TO_DELETE`), as we're modifying/removing fields from the
`ValidatedConfig`. this "felt" like it would inevitably throw compiler
errors as we make `ValidatedConfig` stricter. if/when we decide to get
rid of `d.Config`, we can revisit (and the compiler will make us!)