chore(telemetry): remove info task telemetry call
this commit removes a telemetry call from stencil for the info task.
this call never actually worked, and is something that we don't feel we
want/need to collect.

it also has an additional positive side effect for making the stencil
configuration entity stricter - this call was placed before
loading/validating a configuration (because doing so is not required to
run the info task, making it faster). as `ValidatedConfig` grew in size,
additional fields were added to make a bespoke validated config.
removing this call removes the need to maintain the config used, making
the process of making `ValidatedConfig` stricter slightly easier