refactor(help): make help task require CompilerSystem
with this commit force the help task to accept a CompilerSystem
argument. previously this was made optional when telemetry was added
to the project. in order to not introduce breaking changes to existing
users of the `runtTask` api, `sys` was made optional. now that we know
the property will always be available, the optionality can be safely
removed