chore(run): remove unneeded CompilerSystem checks (#3494)
In 66d0476b (#3079), we added a truthy check for `CompilerSystem` entities.
With the recent move to making that entity required on a `ValidatedConfig` entity
(a6a9171 - #3491), these checks can be safely removed.

This PR removes cases throughout the code base where a truthy check is
required in order to call a function that requires it. In every case, we are able to
utilize the fact that the calling function has an instance of `ValidatedConfig`, which
now has a `sys` on it. 

The original belief was that this would need to be a breaking change (this was
months ago when the ticket was originally written). However, the foundation of this
PR  (a6a9171) does not require this to be a breaking change anymore.