fix(validation): update module location suggestion
this commit fixes the module directory suggestion for the
`dist-custom-elements` output target. in 47c4ccb, we introduced a bug
where we assumed that the output target's `dir` field did not include a
'components' subdirectory. however, this is not the case, as we add
'components' to the `dir` field at the time of validating the output
target. by removing 'components' from the directory in this commit, we
use the value that is stored in `dir` as the source of truth