refactor(validation): move custom-elements check after dist check
this commit moves the check for the `dist-custom-elements` output target
to occur after the `dist` output target check. this results in the
validation for the `module` field in `package.json` giving higher
precedence to the `dist` output target. this change was made for two
reasons:

1. it aligns with the current output of the stencil-component-starter,
   which includes both output targets in its stencil config file
2. it (temporarily) avoids complicating the logic of validating the
   `module` filed when there is more than one output target (by
   restoring the original behavior)

the latter point will be covered in a future task to be completed by the
stencil team. that work was deferred for now in an attempt to avoid
rushing through a design in order to get this bug fix out the door