chore(config): make testing property required field
this commit updates `ValidatedConfig` in order to make the `testing`
field required. this is being done in pursuit of dropping the number of
`strictNullChecks` violations in the code, and was a side task of
stencil-501 (the author of this commit was actively in the testing code,
making it a good opportunity to knock this out).

this commit makes `testing` default to an empty object literal, rather
than `null` or prepopulating existing fields. choosing the former would
require additional optional checks throughout the code (which assumes
`testing` is a defined field, and choosing the latter would expand the
scope of this commit vastly. an empty object literal should incur
minimal runtime overhead, while keeping the size/scope of this commit to
a minimal