chore(config): use ValidatedConfig in output-targets/ (#3496)
this commit propagates the usage of `ValidatedConfig` throughout the
`output-targets/` directory. This is being done following making `sys` a
required field on `ValidatedConfig` in 991843a747fcc165ff13cef9fc888cb4297df271
(#3491), where driving down the number of `sys`-related violations of the compile
time check is the primary driver.  however, this commit does also lay groundwork
for future fields being made required on the validated config, and require less
proliferation of the type in the future.

all output target types were updated in this commit. that includes the legacy angular
output target, and the custom elements bundle output target. both of the aforementioned
targets are intended to be deprecated/removed in stencil v3. however, the level of effort
required to proliferate the change provides us with a more accurate count of
`strictNullChecks` violations, and does not require waiting for a major version bump of
stencil to drive the violation count further