chore(config): migrate additional callsites to ValidatedConfig (#3498)
this commit propagates the usage of `ValidatedConfig` throughout
various parts of the code. the changes are distributed and small in size,
which made it feel unnecessary to split into multiple pull requests, given
the nature of the changes being proposed.

this is being done following making sys a required field on ValidatedConfig
in 991843a (#3491), where driving down the number of sys-related violations of
the compile time check is the primary driver.