fix(compiler): don't break livereload by mangling CSS
This fixes an issue (#3461) where CSS is being inappropriately mangled
when using the hot-reload dev server. The issue has details about
exactly what the problem is, but basically in short if you are using the
`dist-hydrate-script` output target and running the dev server then any
change to a CSS file will cause all styling to be wiped from the related
component in the browser, making for a pretty inadequate developer
experience.

This commit fixes the issue by changing the conditions under which
original CSS selectors are commented out and 'scoped' so that it is only
done when you are have `scoped: true` set on a component (I believe this
was the original intent but for some reason the code wasn't set up to do
this).