chore(github): add dependabot
this commit adds a dependabot configuration to the main stencil project.
it is being added in the hope that we can continue to incrementally
update packages in the compiler, by letting the bot try to incrementally
update dependencies on our behalf.

two packages are ignored in this configuration:
- @types/node - we develop Stencil in Node 16, and do not want later
  versions to be suggested by dependabot
- typescript - v4.8 is not out (it hit release candidate this week), and
  we will be attempting to upgrade to v4.7 in q3

this configuration intentionally only checks the root level manifest
(i.e. it does not check any other `package.json` files in the project).
as we roll this out further, we can expand the usage of the bot. the
root level directory has the highest level of value for us for now

stencil-320: enable dependabot in stencil core