fix(task): consider config sys in task runner (#3518)
this commit updates the generation of a `ValidatedConfig` entity to take
a `sys` property on a user provided `Config` in the context of
`runTask`.

when generating the validated configuration, the `sys` entity takes
precedence over all other configurations, with the belief that in the
event argument was provided, it is the intent of the user to use it. in
the event `sys` is not expressly provided, use the one on the provided
configuration entity. this is useful in the case that a user used
stencil's external facing apis in a flow that:
1. loads a configuration
2. validates that configuration
3. uses that configuration when calling `runTask` directly

this commit fixes a bug introduced making `sys` a required field on
`ValidatedConfig`, where `config.sys` was not properly accounted for.
the bug was initially introduced in
a6a9171a5e8aebd0070e3d6a89e41b1e9199de82 (#3491)