chore(tests): fix null check errors for jest preprocess
this commit removes a handful of strict null check violations that can
be found in stencil's jest preprocessor. most changes are concerned with
making types agree - e.g. if a variable is declared to be of type `T`,
but is assigned by a function that returns `T | U`, the type of the
variable is expanded for agreement.

the two changes that are worth calling out in particular is introducting
a `throw` statement in the event that the `transformOptions` parameter
is falsey both when we process a file and attempt to retrieve a cache
key. from jest 23 thru 27, the function signatures of these two
functions changed a few times (documented in the code today). however,
a typing issue that was introduced in
e0640de27de7aa1b6ebcf96d4298a1c87d9954bb (#3171) was `transformOptions`
could be considered `undefined` by typescript. we know that for jest
[23,26], this value will never be falsey (as it is a part of the call
signature). for jest 27, we re-assign a value to this variable for
backwards compatability reasons.

there is a high level of confidence that this `throw` is safe to
introduce for the following reasons:
1. if `transformOptions` is undefined/null, stencil would already be
   erroring out when running the test task
2. the function signatures changes of these two functions have been
   handled in e0640de27de7aa1b6ebcf96d4298a1c87d9954bb
3. in jest [23,26], `transformOptions` _always_ has a `rootDir` field of
   type `string` (which is the field we're accessing here causing
   errors)