chore(ci): fix dirty git context for browserstack (#3550)
this commit modifies the build_core reusable workflow for stencil's ci
to take into account the event name when checking out the code to be
built. depending on the event name, the HEAD value used to check out
the code can differ. for event's with name 'pull_request_target', HEAD
is the head of the primary branch, in our case, main.

this lead to an issue in the stencil repo where two workflows would run
in order to build/run browserstack:

    build_core, which would use main@HEAD when building the repo, and
    subesquently archiving the results for the next step
    Browserstack Tests, which would checkout the code associated with
    the pull request, the unpack the saved dependencies from main over
    the contents of the pull request

because step 2 would take on files from main when the build archive was
unpacked, we would fail the git status check, because files from main would
clash with files on the checked out commit associated with the pr

this bug was originally introduced in 1cef74f (#3454)