chore(ci): run tech debt burndown on forks (#3552)
this commit changes the event type on the tech debt burndown list. the
pull_request event does not offer the ability run the
'peter-evans/create-or-update-comment' step (see
https://github.com/peter-evans/create-pull-request/blob/main/docs/concepts-guidelines.md#restrictions-on-repository-forks)
for security reasons. in this commit, we use 'pull_request_target' in an
attempt to fix that.

however, this affects what branch is used from in ci from the branch
that the pr was opened with (pull_request), to the `HEAD` of `main`
(pull_request_target). as a result, `actions/checkout` steps must be
modified in order to account for the difference in branch