chore(dev-deps): remove unneeded @types/is-glob (#3549)
this commit removes `@types/is-glob` from the project's root manifest.
`is-glob` does not appear to be used in the project at this time. it
appears it once was in the project, but was removed prior to v1.16.0
(it was removed in 0522e66b5213cb80de72e544ff5e3185f389dd32,
specifically).

this commit is motivated by dependabot attempting to bump the version of
this package in https://github.com/ionic-team/stencil/pull/3547. rather
than bump the version, we should prefer to remove packages we do not use