rework the readme
this commit reworks the readme to remove some verbosity and awkward
language from the document. it's been updated to align a little bit more
with the Ionic Framework's root README file. some content has been
removed with the intent on directing folks to the stencil site (which
we're working on revamping).