chore(deps-dev): remove tslib (#3558)
this commit removes the tslib dependency from the project. the impetus
for this commit and its assoicated pull request was another pull request
created by dependabot to increment the version used in stencil from
v2.1.0 to v2.4.0 (#3543). upon futher investigation, it does not appear
that this library is actually used.

tslib is meant to be used in conjunction with the tsconfig flag
`importHelpers`, which is not present in any `tsconfig.json` file in the
project, nor is it used in any capacity when we validate a loaded config
file. upon further research, it appears we once used the flag in several
tsconfig files, but that ended in
https://github.com/ionic-team/stencil/commit/e4efbf094886502d5a1d76f0912137e09ee3539e#diff-b55cdbef4907b7045f32cc5360d48d262cca5f94062e353089f189f4460039e0.

this commit was validated by building stencil with these changes, as
well as `main@HEAD`, and the output was diff'ed. no changes related to
adding `__assign` or similar helpers was found