chore(compiler): add comments and update types for in-memory FS (#3469)
This updates the in-memory FS file with a bunch of comments explaining
what various methods on the `InMemoryFileSystem` object do.

It also refactors where the relevant types live. Instead of having a
separate interface declared outside of the file where the
implementation is, we instead derive the `InMemoryFileSystem` from the
return value of the function which creates the object. This lets us have
comments annotating the implementation code which will also show up as
type hints throughout the codebase.