feat(typescript): add support for typescript v4.7 (#3530)
this pr adds support through v4.7.4 of typescript. to support 
v4.6 and v4.7, minor changes were required for each:

- to add support for TypeScript 4.6, the `lib` fields of a few `tsconfig.json` files needed to
be updated to continue to support `Array.prototype.flat()`. TS no longer bundles 
`flat` in `esnext.array`, but does in `es2019`. as a result, the former value was removed,
while `es2017` was bumped to `es2019`
- to add support for TypeScript v4.7, `Performance` gained the `EventCount` field, which
was added to `MockPerformance` to adhere to that interface.

STENCIL-435: Upgrade TypeScript to v4.7