fix(compiler): add config flags for transpiling path aliases (#3559)
* fix(compiler): add config flags for transpiling path aliases

This commit adds config flags to enable the transpilation of import path aliases as a part of the `dist-collection` output target. Part of the work for STENCIL-437

* fix(compiler): update terminology from 'transpile' to 'transform'

This commit updates the language for the config option of `dist-collection` from `transpileAliasedImportPaths` to `transformAliasedImportPaths`. This also adds validation logic and test coverage for the `validateCollection` function

* fix(compiler): use `@example` JSdoc tag for code formatting

* fix(compiler): test cleanup for `dist` targets

This commit cleans up some test code for the `dist` and `dist-collection` targets

* fix(compiler): fix overwrite race condition for `dist-collection`

This commit fixes an overwrite race condition for duplicate `dist-collection` output targets writing to the same file location(s).

* test(compiler): fix transform test normalization?

* fix(compiler): use `normalizePath` util in path alias transformer