chore(types): remove type-check errors from `spec-page`
this commit updates the typings for `spec-page.ts` to remove
`strictNullChecks` violations from the aforementioned file. the private
type `CompnentTestingConstructor` has some of its fields widened to
accepted `null` to reflect the state of the the existing assignments.
similarly, optional chaining is added to the conditionals to prevent
assignments/accesses that would otherwise be illegal