refactor(decorators): fix `strictNullCheck` violation
this commit fixes a single `strictNullChecks` violation where the result
of an expression could set a local variable to be of type `string |
false`, where `false` was not an accepted value on `Set.prototype.has`.
this change ought to be backwards compatible, as no `Set` provided to
this function includes `false` (or any value that is coerced to
`false`).

the function signature was updated to include a return type, and the
function name was updated to provide a better description of what the
function does.

the deprecated call to `createNodeArray` has been moved to the
`ts.factory` for general code health.

during this exercise, it was discovered there was an unused instance of
`removeDecorators` in the project's utilities. it has been removed as a
part of this commit