chore(types): remove type-check errors from spec-page  (#3587)
this commit updates the typings for `spec-page.ts` to remove
`strictNullChecks` violations from the aforementioned file. the private
type `CompnentTestingConstructor` has some of its fields widened to
accepted `null` to reflect the state of the the existing assignments.
similarly, optional chaining is added to the conditionals to prevent
assignments/accesses that would otherwise be illegal

this commit also updates the function signature for `getHostRef` in the
testing package, to accept an argument of type undefined. today, it is
possible for the function to accept a value that is `undefined`; this
change only updates the type system to reflect that fact.