chore(repo): format additional files in repo (#3598)
this commit updates our prettier configuration to format all yaml files
in the `.github` directory at the root of the the repo. both `.yml` and
`.yaml` were added. although the latter isn't used today, it is added as
a bit of 'future proofing'.

with the updated config, prettier was run on the repo. the only change
that is worth noting is that none of the double to single quote changes
_shouldn't_ be of issue, as the contents of the strings themselves don't
have an double quotes in them (e.g. `""something""`)

this commit updates the prettier configuration for the project to format
all top level .js files. this commit then formats those files.