chore(lint): add auto-sort for imports/exports (#3583)
this commit adds the 'simple-import-sort' eslint plugin package
to the repository, and sorts imports according to its default
heuristic. this is being done to promote consistency within the
repo, as we gradually build out our list of lint rules.

it also updates the order of ci steps, since this plugin can create
prettier/format errors when it reorders imports.