chore(compiler): add JSDocs and some types related to @State decorator
This updates the docstrings and typings for a few functions which have
to do with the `@State` decorator and, in particular, the code which
is used to convert class field declarations to getters.

This also switches over to use `ts.factory` methods in a few places
where deprecation messages were popping up.