chore(dev-deps): upgrade eslint dependencies (#3621)
this commit updates our existing eslint dependencies to the latest and
greateest version of each. as a part of this upgrade, the jest plugin
for eslint was incremented by one major version. as a result, there was
one breaking change where alias matches are not to be used anymore.
those have been updated in this commit as well