chore(dev-deps): upgrade release script dependencies (#3622)
this commit increments the version of the semver package that is used as
a part of the release process for stencil. this is being done in order
to attempt to keep the dependencies for the project up to date.

this commit increments the version of the typings used for listr, a
package used to help release stencil. this is done in an effort to keep
dependencies for the project up to date, while grouping them for a
single pull request by the area of the code they belong in (in this
case, release scripts)

this commit updates the 'open' dependency to use the latest and greatest
version of the package. this is being done in order to keep the
dependencies of stencil up to date, while simultaneously attempting to
group the updates by the section of the code that they belong in (in
this case, release scripts)