chore(dev-deps): remove unused dependencies
this commit removes everal dependencies that are not used in stencil
today. they are, in alphabetical order:
- `@types/autoprefixer`
- `@types/mime-types`
- `@types/autoprefixer`
- `core-js-builder`
- `css`
- `fast-deep-equal`

these packages are being removed due to no usage of them being found
within stencil.

the `license.ts` file was updated accordingly (removing entries that
were removed), and `NOTICE.md` was regenerated using the updates in the
license file