refactor(ci): pull git context out to separate action
this commit pulls the check of the git context out to a separate action.
this allows us to use it in other places than only the browserstack job.
it has been added to the e2e test suite, which also may generate a new
`components.d.ts` file as a result of changes to the compiler, then
building the tests in `end-to-end`