chore(ci): fix build out-of-memory errors
this commit introduces an `.npmrc` file to the project that specifies a
new value for `--max-old-space-size`. by default, node allocates 2GB of
memory to a process. it is possible for stencil builds to near this
limit, causing an out of memory failure. we have seen this more often in
continuous integration (github actions) than anywhere else.

an `.npmrc` file was chosen here as an operating system, shell, and
build-agnostic solution. that is, this solution is intended to work:
- on linux, macos, and windows (although we only use linux for building,
  and the team develops solely on macos)
- on any shell (no hacks to the run command either in `package.json` or
  in the ci job for building the project)
- we use the same value in ci and locally (machine configuraton parity)

the value of '4096' was chosen for this configuration field by doubling
the existing default of '2048'