refactor(types): resolve strictNullChecks in `find-config`
this commit resolves a type error when finding a configuration. it is a
common paradigm in stencil to declare a 'result' object at the beginning
of a function. however this presents difficulty when a the fields of the
object's type are strictly typed (i.e. don't permit null values). that
is the case for this file, where the `FindConfigResults` type does not
permit a null config path. however, procedurally, we _always_ set this
value. to circumvent the type issues, move the declaration of the result
object to be right before its first meaningful use.