refactor(types): drop strictNullChecks count in cli (#3632)
this commit is the amalgamation of several commits

refactor(types): explicitly handle null task

this commit explicitly handles the the case of a null task, which we
have decided to allow on a type-level. a new test case is added to
verify the new behavior: when the null task is found, we show the help
prompt


refactor(types): resolve strictNullChecks in `find-config`

this commit resolves a type error when finding a configuration. it is a
common paradigm in stencil to declare a 'result' object at the beginning
of a function. however this presents difficulty when a the fields of the
object's type are strictly typed (i.e. don't permit null values). that
is the case for this file, where the `FindConfigResults` type does not
permit a null config path. however, procedurally, we _always_ set this
value. to circumvent the type issues, move the declaration of the result
object to be right before its first meaningful use.


refactor(tests): update telemetry helper tests

this commit cleans up the telemetry helper tests and the source code
they exercise by removing strict null check violations and improving the
coverage for the case where flags we use may not be a boolean value


refactor(types): widen telemetry types

this commit widens various telemetry types to better reflect real
world scenarios/align better with the rest of the type system. most
of the issues resolved in this commit were introduced alongside the
feature.


refactor(types): type safety in ionic config

this commit makes the reading of telemetry tokens a little type safer by
first checking if such a token exists before running it through the
`test()`. perhaps this was a premature optimization, but this was
preferred over nullish coalescing the property read inside of `test()`
to avoid invoking the regex engine